still need to refactor the API (have moved part into main image section), create a redux file, and update the component (to use the dispatch)
