+++ .gitignore
@@ -2,3 +2,4 @@ node_modules
 bundle.js
 bundle.js.map
 secrets.js
+public/uploads/
